Add files via upload
Technologies Used
HTML: Used for structuring the content and layout of the web pages.
CSS: Utilized for styling the elements and applying visual effects.

Acknowledgments
This project was inspired by the user interface of YouTube and was created for educational purposes only. It does not aim to replicate any backend functionality or infringe on YouTube's intellectual property rights.